requirements.txt
Flask
rembg
Pillow
gunicorn